Updating Common Ports
I was working with a customer to activate a S3 File storage gateway and it was failing. When looking at the traffic flow from their database it is because the storage gateway is using the ports 1026,1027,1028, and 1031 in addition to 443. This is documented in the FSx Gateway page but is listed as being necessary for the FSx Gateway and not as being common across all gateway types. This change is to move the ports to the common ports table since they appear to be common.